# swagger/swagger.yaml
consumes:
- application/json
- application/xml
- application/gob
- application/x-gob
definitions:
  GoaMyexampleBottle:
    description: A bottle of wine (default view)
    example:
      href: Voluptates non excepturi.
      id: 6.9965735e+18
      name: Tenetur et saepe ratione debitis enim sapiente.
    properties:
      href:
        description: API href for making requests on the bottle
        example: Voluptates non excepturi.
        type: string
      id:
        description: Unique bottle ID
        example: 6.9965735e+18
        format: int64
        type: integer
      name:
        description: Name of wine
        example: Tenetur et saepe ratione debitis enim sapiente.
        type: string
    required:
    - id
    - href
    - name
    title: 'Mediatype identifier: application/vnd.goa.myexample.bottle+json'
    type: object
  error:
    description: Error response media type (default view)
    example:
      code: invalid_value
      detail: Value of ID must be an integer
      id: 3F1FKVRR
      meta:
        timestamp: 1.458609e+09
      status: "400"
    properties:
      code:
        description: an application-specific error code, expressed as a string value.
        example: invalid_value
        type: string
      detail:
        description: a human-readable explanation specific to this occurrence of the
          problem.
        example: Value of ID must be an integer
        type: string
      id:
        description: a unique identifier for this particular occurrence of the problem.
        example: 3F1FKVRR
        type: string
      meta:
        additionalProperties: true
        description: a meta object containing non-standard meta-information about
          the error.
        example:
          timestamp: 1.458609e+09
        type: object
      status:
        description: the HTTP status code applicable to this problem, expressed as
          a string value.
        example: "400"
        type: string
    title: 'Mediatype identifier: application/vnd.goa.error'
    type: object
host: localhost:8080
info:
  description: A simple goa service
  title: The virtual wine cellar
  version: ""
paths:
  /bottles/{bottleID}:
    get:
      description: 指定されたIDのボトルを取得する
      operationId: bottle#show
      parameters:
      - description: Bottle ID
        in: path
        name: bottleID
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GoaMyexampleBottle'
        "404":
          description: Not Found
      schemes:
      - http
      summary: show bottle
      tags:
      - bottle
  /swagger.json:
    get:
      operationId: swagger#/swagger.json
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
      schemes:
      - http
      summary: Download swagger/swagger.json
  /swaggerui/{filepath}:
    get:
      operationId: swagger#/swaggerui/*filepath
      parameters:
      - description: Relative file path
        in: path
        name: filepath
        required: true
        type: string
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
        "404":
          description: File not found
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: Download swaggerui/dist
produces:
- application/json
- application/xml
- application/gob
- application/x-gob
responses:
  NotFound:
    description: Not Found
  OK:
    description: OK
    schema:
      $ref: '#/definitions/GoaMyexampleBottle'
schemes:
- http
swagger: "2.0"